requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
python-jose[cryptography]
httpx
pydantic-settings
pydantic[email]
passlib[bcrypt]==1.7.4